+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/Implementation/EntityFrameworkDiagnosticListener.cs
@@ -19,6 +19,7 @@ internal sealed class EntityFrameworkDiagnosticListener : ListenerHandler
 
     internal const string AttributePeerService = "peer.service";
     internal const string AttributeServerAddress = "server.address";
+    internal const string AttributeServerPort = "server.port";
     internal const string AttributeDbSystem = "db.system";
     internal const string AttributeDbName = "db.name";
     internal const string AttributeDbNamespace = "db.namespace";
@@ -40,6 +41,8 @@ internal sealed class EntityFrameworkDiagnosticListener : ListenerHandler
     private readonly PropertyFetcher<CommandType> commandTypeFetcher = new("CommandType");
     private readonly PropertyFetcher<string> commandTextFetcher = new("CommandText");
     private readonly PropertyFetcher<Exception> exceptionFetcher = new("Exception");
+    private readonly PropertyFetcher<string> hostFetcher = new("Host");
+    private readonly PropertyFetcher<string> portFetcher = new("Port");
 
     private readonly EntityFrameworkInstrumentationOptions options;
 
@@ -141,10 +144,24 @@ internal sealed class EntityFrameworkDiagnosticListener : ListenerHandler
                                 break;
                         }
 
-                        var dataSource = (string)this.dataSourceFetcher.Fetch(connection);
-                        if (!string.IsNullOrEmpty(dataSource))
+                        var host = (string)this.hostFetcher.Fetch(connection);
+                        if (!string.IsNullOrEmpty(host))
                         {
-                            activity.AddTag(AttributeServerAddress, dataSource);
+                            activity.AddTag(AttributeServerAddress, host);
+                        }
+                        else
+                        {
+                            var dataSource = (string)this.dataSourceFetcher.Fetch(connection);
+                            if (!string.IsNullOrEmpty(dataSource))
+                            {
+                                activity.AddTag(AttributeServerAddress, dataSource);
+                            }
+                        }
+
+                        var port = (string)this.portFetcher.Fetch(connection);
+                        if (!string.IsNullOrEmpty(port))
+                        {
+                            activity.AddTag(AttributeServerPort, port);
                         }
 
                         if (this.options.EmitOldAttributes)