+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 * Added case when `providerName` has 
-value `"Devart.Data.Oracle.Entity.EFCore"` ([#2465](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2465))
+  value `Devart.Data.Oracle.Entity.EFCore`.
+  ([#2465](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2465))
 
 ## 1.10.0-beta.1