+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Added case when `providerName` has 
+* Added case when `providerName` has
   value `Devart.Data.Oracle.Entity.EFCore`.
   ([#2465](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2465))