+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Added code to fix exception thrown when Devart Oracle is used
+(`this.commandTypeFetcher.Fetch(command)`). If Devart Oracle is used,
+the SQL statement will be extracted from the payload.
+([#2466](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2466))
+
 ## 1.10.0-beta.1
 
 Released 2024-Dec-09