+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-* Fixed the `server.address` field in EntityFrameworkCore spans to populate with the server domain name without the `tcp` prefix.
+* Fixed the `server.address` field in EntityFrameworkCore spans to populate
+  with the server domain name without the `tcp` prefix.
   ([#IssueNumber](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/issues/IssueNumber))
 
 ## 1.10.0-beta.1