+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Updated OpenTelemetry core component version(s) to `1.11.0`.
   ([#2470](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2470))
 
+* Attribute `db.system` reports `oracle` when
+  `Devart.Data.Oracle.Entity.EFCore` is used a provider.
+  ([#2465](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2465))
+
 ## 1.10.0-beta.1
 
 Released 2024-Dec-09