+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -9,6 +9,10 @@
   `Devart.Data.Oracle.Entity.EFCore` is used a provider.
   ([#2465](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2465))
 
+* Correctly report `db.statement`/`db.query.text` when
+  Devart Oracle is used as a provider.
+  ([#2466](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2466))
+
 ## 1.10.0-beta.1
 
 Released 2024-Dec-09